+++ cmd/template_pull.go
@@ -14,14 +14,15 @@ var templatePullCmd = &cobra.Command{
 	Example: `To pull the official templates, run:
 
     $ safira template pull`,
-	RunE:  runTemplatePull,
+	RunE:                       runTemplatePull,
+	SuggestionsMinimumDistance: 1,
 }
 
 func init() {
 	templateCmd.AddCommand(templatePullCmd)
 }
 
-func runTemplatePull(cmd *cobra.Command, args []string) error {
+func runTemplatePull(cmd *cobra.Command, _ []string) error {
 	verboseFlag, _ := cmd.Flags().GetBool("verbose")
 	if err := get.DownloadTemplate(faasTemplateRepo, true, verboseFlag); err != nil {
 		return err